refactor(repo): reorganize folder structure to add vhdl and systemverilog files in different directories
