+++ pytest.ini
@@ -2,4 +2,4 @@
 minversion = 6.0
 addopts = -ra -q -v --tb=short
 testpaths =
-    src/bench
+    src/bench/cocotb
