+++ src/bench/cocotb/ascon/ascon_model.py
