+++ src/bench/cocotb/diffusion_layer/test_diffusion_layer.py
@@ -1,16 +1,15 @@
 """
-Testbench for the XOR Begin Layer.
+Testbench for the Diffusion Layer module.
 
-This module tests the XOR Begin Layer function module by comparing the
-output of the Python implementation with the Verilog implementation.
+This module tests the Diffusion Layer module by comparing the
+output of the Python implementation with the VHDL implementation.
 
-Author: Timothée Charrier
+@author: Timothée Charrier
 """
 
 from __future__ import annotations
 
 import os
-import random
 import sys
 from pathlib import Path
 from typing import TYPE_CHECKING
@@ -18,10 +17,10 @@ from typing import TYPE_CHECKING
 import cocotb
 from cocotb.triggers import Timer
 from cocotb_tools.runner import get_runner
-from xor_begin_model import XorBeginModel
+from diffusion_layer_model import DiffusionLayerModel
 
 # Add the directory containing the utils.py file to the Python path
-sys.path.insert(0, str(object=(Path(__file__).parent.parent.parent).resolve()))
+sys.path.insert(0, str(object=(Path(__file__).parent.parent).resolve()))
 
 from cocotb_utils import (
     generate_coverage_report_questa,
@@ -34,13 +33,8 @@ if TYPE_CHECKING:
     from cocotb.handle import HierarchyObject
     from cocotb_tools.runner import Runner
 
-
 INIT_INPUTS = {
     "i_state": init_hierarchy(dims=(5,), bitwidth=64, use_random=False),
-    "i_data": 0,
-    "i_key": 0,
-    "i_enable_xor_key": 0,
-    "i_enable_xor_data": 0,
 }
 
 
@@ -81,18 +75,18 @@ async def reset_dut_test(dut: HierarchyObject) -> None:
     """
     try:
         # Define the model
-        xor_begin_model = XorBeginModel()
+        diffusion_layer_model = DiffusionLayerModel()
 
         # Initialize the DUT
         await initialize_dut(dut=dut, inputs=INIT_INPUTS)
 
         # Verify the output
-        xor_begin_model.assert_output(dut=dut, inputs=INIT_INPUTS)
+        diffusion_layer_model.assert_output(dut=dut, state=INIT_INPUTS["i_state"])
 
     except Exception as e:
         dut_state = get_dut_state(dut=dut)
         formatted_dut_state: str = "\n".join(
-            f"{key}: {value}" for key, value in dut_state.items()
+            [f"{key}: {value}" for key, value in dut_state.items()],
         )
         error_message: str = (
             f"Failed in reset_dut_test with error: {e}\n"
@@ -103,11 +97,9 @@ async def reset_dut_test(dut: HierarchyObject) -> None:
 
 
 @cocotb.test()
-async def xor_begin_test(dut: HierarchyObject) -> None:
+async def diffusion_layer_test(dut: HierarchyObject) -> None:
     """
-    Test the DUT's behavior during normal computation.
-
-    Verifies that the output is correctly computed.
+    Test the Diffusion Layer module.
 
     Parameters
     ----------
@@ -122,73 +114,59 @@ async def xor_begin_test(dut: HierarchyObject) -> None:
     """
     try:
         # Define the model
-        xor_begin_model = XorBeginModel()
+        diffusion_layer_model = DiffusionLayerModel()
 
-        # Reset the DUT
         await reset_dut_test(dut=dut)
 
         # Test with specific inputs
-        dut_inputs = [
+        dut_inputs: list[dict[str, list[int]]] = [
             {
-                "i_state": [0xFFFFFFFFFFFFFFFF] * 5,
-                "i_data": 0xFFFFFFFFFFFFFFFF,
-                "i_key": 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
-                "i_enable_xor_key": 0,
-                "i_enable_xor_data": 0,
+                "i_state": [
+                    0x8849060F0C0D0EFF,
+                    0x80410E05040506F7,
+                    0xFFFFFFFFFFFFFF0F,
+                    0x80400406000000F0,
+                    0x0808080A08080808,
+                ],
             },
             {
-                "i_state": [0xFFFFFFFFFFFFFFFF] * 5,
-                "i_data": 0xFFFFFFFFFFFFFFFF,
-                "i_key": 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
-                "i_enable_xor_key": 1,
-                "i_enable_xor_data": 0,
+                "i_state": [
+                    0x8CBD402180B4D43D,
+                    0x778B87B53BCCBF49,
+                    0x3E7883FEF208E8C0,
+                    0x0B48487C6AFB2C4D,
+                    0x0F20CDA96AE53627,
+                ],
             },
             {
-                "i_state": [0xFFFFFFFFFFFFFFFF] * 5,
-                "i_data": 0xFFFFFFFFFFFFFFFF,
-                "i_key": 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
-                "i_enable_xor_key": 0,
-                "i_enable_xor_data": 1,
-            },
-            {
-                "i_state": [0xFFFFFFFFFFFFFFFF] * 5,
-                "i_data": 0xFFFFFFFFFFFFFFFF,
-                "i_key": 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,
-                "i_enable_xor_key": 1,
-                "i_enable_xor_data": 1,
+                "i_state": [
+                    0xBAF6B13AFEB21E28,
+                    0xABA64F6758F07EB1,
+                    0x59D013C5A157E2F3,
+                    0xA4CDCEB4CF026350,
+                    0xA982986B3A1FBA70,
+                ],
             },
         ]
 
         for inputs in dut_inputs:
             await initialize_dut(dut=dut, inputs=inputs)
-            xor_begin_model.assert_output(dut=dut, inputs=inputs)
-
-        # Test with random inputs
-        for _ in range(10):
-            random_inputs = {
-                "i_state": init_hierarchy(dims=(5,), bitwidth=64, use_random=True),
-                "i_data": random.randint(0, 2**64 - 1),
-                "i_key": random.randint(0, 2**128 - 1),
-                "i_enable_xor_key": random.randint(0, 1),
-                "i_enable_xor_data": random.randint(0, 1),
-            }
-            await initialize_dut(dut=dut, inputs=random_inputs)
-            xor_begin_model.assert_output(dut=dut, inputs=random_inputs)
+            diffusion_layer_model.assert_output(dut=dut, state=inputs["i_state"])
 
     except Exception as e:
-        dut_state: dict = get_dut_state(dut=dut)
+        dut_state = get_dut_state(dut=dut)
         formatted_dut_state: str = "\n".join(
-            f"{key}: {value}" for key, value in dut_state.items()
+            [f"{key}: {value}" for key, value in dut_state.items()],
         )
         error_message: str = (
-            f"Failed in xor_begin_test with error: {e}\n"
+            f"Failed in diffusion_layer_test with error: {e}\n"
             f"DUT state at error:\n"
             f"{formatted_dut_state}"
         )
         raise RuntimeError(error_message) from e
 
 
-def test_xor_begin() -> None:
+def test_diffusion_layer() -> None:
     """
     Function Invoked by the test runner to execute the tests.
 
@@ -203,13 +181,13 @@ def test_xor_begin() -> None:
 
     # Define the top-level library and entity
     library: str = "lib_rtl"
-    entity: str = "xor_begin"
+    entity: str = "diffusion_layer"
 
     # Default Generics Configuration
     generics: dict[str, str] = {}
 
     # Define paths
-    rtl_path: Path = (Path(__file__).parent.parent.parent.parent / "rtl/").resolve()
+    rtl_path: Path = Path(__file__).parent.parent.parent.parent / "rtl" / "verilog"
     build_dir: Path = Path("sim_build")
 
     # Define the coverage file and output folder
@@ -224,7 +202,7 @@ def test_xor_begin() -> None:
     # Define the sources
     sources: list[str] = [
         f"{rtl_path}/ascon_pkg.sv",
-        f"{rtl_path}/xor/xor_begin.sv",
+        f"{rtl_path}/diffusion_layer/diffusion_layer.sv",
     ]
 
     # Define the build and test arguments
@@ -248,6 +226,8 @@ def test_xor_begin() -> None:
             "0",
             "-Wall",
             "--coverage",
+            "--coverage-max-width",
+            "320",
         ]
         test_args: list[str] = []
         pre_cmd = None
@@ -308,4 +288,4 @@ def test_xor_begin() -> None:
 
 
 if __name__ == "__main__":
-    test_xor_begin()
+    test_diffusion_layer()
