+++ src/bench/cocotb/example/adder/test_adder.py
@@ -85,7 +85,9 @@ def test_counter_runner() -> None:
     generics: dict[str, str] = {"G_DATA_WIDTH": 8}
 
     # Define paths
-    rtl_path: Path = (Path(__file__).parent.parent.parent.parent / "rtl/").resolve()
+    rtl_path: Path = (
+        Path(__file__).parent.parent.parent.parent.parent / "rtl" / "verilog"
+    )
     build_dir: Path = Path("sim_build")
 
     # Define the coverage file and output folder
