+++ src/bench/cocotb/substitution_layer/sbox/sbox_model.py
