+++ src/bench/cocotb/substitution_layer/sbox/test_sbox.py
@@ -197,7 +197,9 @@ def test_sbox() -> None:
     generics: dict[str, str] = {}
 
     # Define paths
-    rtl_path: Path = (Path(__file__).parent.parent.parent.parent / "rtl/").resolve()
+    rtl_path: Path = (
+        Path(__file__).parent.parent.parent.parent.parent / "rtl" / "verilog"
+    )
     build_dir: Path = Path("sim_build")
 
     # Define the coverage file and output folder
