+++ src/bench/cocotb/xor/xor_begin/xor_begin_model.py
