+++ src/rtl/systemverilog/example/counter.sv
